+++ et-common-libs/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-15710] Baseline implementation for Accelerator lib docs.
 - [SW-15710] Adding a docs prototype with examples
 - [SW-15805] Adding doxygen docs support for memory operations API with examples
+- [SW-15803] Adding doxygen docs support for tensor operations API 
 ### Changed
 ### Deprecated
 ### Removed